@@ -1,6 +1,6 @@
 module github.com/aws/karpenter-core
 
-go 1.20
+go 1.21
 
 require (
 	github.com/Pallinder/go-randomdata v1.2.0
