@@ -1,5 +1,4 @@
 //go:build !ignore_autogenerated
-// +build !ignore_autogenerated
 
 /*
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -69,7 +68,8 @@ func (in *StateNode) DeepCopyInto(out *StateNode) {
 			if val == nil {
 				(*out)[key] = nil
 			} else {
-				in, out := &val, &outVal
+				inVal := (*in)[key]
+				in, out := &inVal, &outVal
 				*out = make(v1.ResourceList, len(*in))
 				for key, val := range *in {
 					(*out)[key] = val.DeepCopy()
@@ -86,7 +86,8 @@ func (in *StateNode) DeepCopyInto(out *StateNode) {
 			if val == nil {
 				(*out)[key] = nil
 			} else {
-				in, out := &val, &outVal
+				inVal := (*in)[key]
+				in, out := &inVal, &outVal
 				*out = make(v1.ResourceList, len(*in))
 				for key, val := range *in {
 					(*out)[key] = val.DeepCopy()
@@ -103,7 +104,8 @@ func (in *StateNode) DeepCopyInto(out *StateNode) {
 			if val == nil {
 				(*out)[key] = nil
 			} else {
-				in, out := &val, &outVal
+				inVal := (*in)[key]
+				in, out := &inVal, &outVal
 				*out = make(v1.ResourceList, len(*in))
 				for key, val := range *in {
 					(*out)[key] = val.DeepCopy()
@@ -120,7 +122,8 @@ func (in *StateNode) DeepCopyInto(out *StateNode) {
 			if val == nil {
 				(*out)[key] = nil
 			} else {
-				in, out := &val, &outVal
+				inVal := (*in)[key]
+				in, out := &inVal, &outVal
 				*out = make(v1.ResourceList, len(*in))
 				for key, val := range *in {
 					(*out)[key] = val.DeepCopy()
