@@ -1,5 +1,4 @@
 //go:build !ignore_autogenerated
-// +build !ignore_autogenerated
 
 /*
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -56,7 +55,8 @@ func (in *HostPortUsage) DeepCopyInto(out *HostPortUsage) {
 			if val == nil {
 				(*out)[key] = nil
 			} else {
-				in, out := &val, &outVal
+				inVal := (*in)[key]
+				in, out := &inVal, &outVal
 				*out = make([]HostPort, len(*in))
 				for i := range *in {
 					(*in)[i].DeepCopyInto(&(*out)[i])
@@ -88,7 +88,8 @@ func (in *VolumeUsage) DeepCopyInto(out *VolumeUsage) {
 			if val == nil {
 				(*out)[key] = nil
 			} else {
-				in, out := &val, &outVal
+				inVal := (*in)[key]
+				in, out := &inVal, &outVal
 				*out = make(sets.Set[string], len(*in))
 				for key, val := range *in {
 					(*out)[key] = val
@@ -105,14 +106,16 @@ func (in *VolumeUsage) DeepCopyInto(out *VolumeUsage) {
 			if val == nil {
 				(*out)[key] = nil
 			} else {
-				in, out := &val, &outVal
+				inVal := (*in)[key]
+				in, out := &inVal, &outVal
 				*out = make(Volumes, len(*in))
 				for key, val := range *in {
 					var outVal map[string]sets.Empty
 					if val == nil {
 						(*out)[key] = nil
 					} else {
-						in, out := &val, &outVal
+						inVal := (*in)[key]
+						in, out := &inVal, &outVal
 						*out = make(sets.Set[string], len(*in))
 						for key, val := range *in {
 							(*out)[key] = val
@@ -153,7 +156,8 @@ func (in Volumes) DeepCopyInto(out *Volumes) {
 			if val == nil {
 				(*out)[key] = nil
 			} else {
-				in, out := &val, &outVal
+				inVal := (*in)[key]
+				in, out := &inVal, &outVal
 				*out = make(sets.Set[string], len(*in))
 				for key, val := range *in {
 					(*out)[key] = val
